+++ frontend/models/service.js
@@ -4,21 +4,25 @@ var messages = require('../messages');
 var _base = require('./_base');
 var servicearea = require('./servicearea');
 var servicetype = require('./servicetype');
+var provider = require('./provider');
+var api = require('../api');
 
 
 var Service = _base.BaseModel.extend({
     apiname: 'services',
     loadSubModels: function () {
         var self = this;
-        var area = new servicearea.ServiceArea({url: this.get('area_of_service')});
-        var type = new servicetype.ServiceType({url: this.get('type')});
+        var provider_fetch_url = api.getAbsoluteAPIURL(this.get('provider_fetch_url'));
+        var provider_fetch = provider.fetchAndCache(provider_fetch_url);
 
-        var wait = [area.fetch(), type.fetch()];
+        var wait = [_base.preloaded, provider_fetch];
 
         return new Promise(function(resolve, error) {
-            Promise.all(wait).then(function(){
-                self.attributes.servicearea = area;
-                self.attributes.servicetype = type;
+            Promise.all(wait).then(function(results){
+                var preloaded = results[0];
+                self.attributes.servicearea = preloaded.servicearea.getByUrl(self.get('area_of_service'));
+                self.attributes.servicetype = preloaded.servicetype.getByUrl(self.get('type'));;
+                self.attributes.provider = provider_fetch;
 
                 resolve(self);
             }, function onerror(error) {
@@ -32,6 +36,22 @@ var Service = _base.BaseModel.extend({
         var data = _base.BaseModel.prototype.data.apply(this, arguments);
         data.isApproved = this.isApproved();
         data.isRejected = this.isRejected();
+        window.service = this;
+        if (typeof this.attributes.servicearea !== 'undefined') {
+            if (typeof this.attributes.servicearea.data === 'function') {
+                data.servicearea = this.attributes.servicearea.data();
+            }
+        }
+        if (typeof this.attributes.servicetype !== 'undefined') {
+            if (typeof this.attributes.servicetype.data === 'function') {
+                data.servicetype = this.attributes.servicetype.data();
+            }
+        }
+        if (typeof this.attributes.provider !== 'undefined') {
+            if (typeof this.attributes.provider.data === 'function') {
+                data.provider = this.attributes.provider.data();
+            }
+        }
         return data;
     },