+++ frontend/router.js
@@ -12,14 +12,18 @@ var views = {
     "register-changed": require('./views/provider-form-changed'),
     "account-activate": require('./views/account-activate'),
     "service": require('./views/service-form'),
-    "feedback": require('./views/feedback'),
+    "feedback-form": require('./views/feedback-form'),
+    "feedback-confirm": require('./views/feedback-confirm'),
+    "feedback-help": require('./views/feedback-help'),
     "map": require('./views/map'),
+    "search-list": require('./views/search-list'),
     "service-cancel": require('./views/service-cancel'),
     "service-list": require('./views/service-list'),
+    "service-detail": require('./views/service-detail'),
     "login": require('./views/login'),
     "password-reset": require('./views/password-reset'),
     "password-reset-form": require('./views/password-reset-form'),
-    "admin": require('./views/admin')
+    "admin": require('./views/admin'),
 };
 
 var view, viewName;
@@ -59,6 +63,10 @@ function loadPage(name, params) {
 }
 
 module.exports = Backbone.Router.extend({
+    initialize: function() {
+        this.route(/search\/?/, loadPage("search-list"));
+        this.route(/search\/map/, loadPage("map"));
+    },
     routes: {
         "": function() {
             if (config.get('forever.authToken')) {
@@ -72,10 +80,17 @@ module.exports = Backbone.Router.extend({
         "register/changed": loadPage("register-changed"),
         "register/confirm": loadPage("register-confirm"),
         "register/verify/:key": loadPage("account-activate", ['key']),
-        "service": loadPage("service"),
-        "service/:id": loadPage("service", ['id']),
-        "feedback": loadPage("feedback"),
-        "search/map": loadPage("map"),
+        "manage/service": loadPage("service"),
+        "manage/service/:id": loadPage("service", ['id']),
+        "manage/service/cancel/:id": loadPage("service-cancel", ['id']),
+        "manage/service-list": loadPage("service-list"),
+
+        "service/:id": loadPage("service-detail", ['id']),
+
+        "feedback": loadPage("feedback-help"),
+        "feedback/confirm": loadPage("feedback-confirm"),
+        "feedback/:id": loadPage("feedback-form", ['id']),
+
         "service/cancel/:id": loadPage("service-cancel", ['id']),
         "service-list": loadPage("service-list"),
         "login": loadPage("login"),