+++ frontend/views/provider-form.js
@@ -74,7 +74,7 @@ module.exports = Backbone.View.extend({
     events: {
         "click .form-btn-submit": function() {
             var $el = this.$el;
-            var data = forms.collect($el);
+            var data = forms.collect($el, this.provider);
             var is_new = this.provider === undefined;
 
             var $submit = $el.find('.form-btn-submit');
@@ -117,7 +117,7 @@ module.exports = Backbone.View.extend({
                     },
                     function error(errors) {
                         $submit.removeAttr('disabled');
-                        messages.error(errors);
+                        // forms.js has already displayed any errors
                     }
                 );