+++ frontend/views/service-form.js
@@ -1,5 +1,6 @@
 var Backbone = require('backbone'),
 api = require('../api'),
+config = require('../config'),
 template = require("../templates/service-form.hbs"),
 i18n = require('i18next-client'),
 forms = require('../forms'),
@@ -92,6 +93,7 @@ module.exports = Backbone.View.extend({
 
     events: {
         "click button.remove": function(ev) {
+            ev.preventDefault();
             var $btn = $(ev.target);
             var $row = $btn.closest('.criteria');
             var criteriaCount = $('.criteria').length;
@@ -103,6 +105,7 @@ module.exports = Backbone.View.extend({
             return false;
         },
         "click button.add": function(ev) {
+            ev.preventDefault();
             var $btn = $(ev.target);
             var $row = $btn.closest('.criteria');
             var $newRow = $row.clone();
@@ -128,20 +131,34 @@ module.exports = Backbone.View.extend({
 
             return false;
         },
-        "click .form-btn-submit": function() {
+        "click .form-btn-submit": function(e) {
+            e.preventDefault();
             messages.clear();
             var $el = this.$el;
-            var data = forms.collect($el);
+            var update_of = this.update_of;
+            var data = forms.collect($el, update_of);
             // Like form fields, omit empty values from the data
             remove_empty_fields(data);
             // change criteria items from strings to dictionaries,
             // omitting blank ones
-            var i, name, criteria = [];
+            var i, name, criteria = [], criteria_data, criteria_id, previous_criteria = {};
             var criteria_length = !!data.selection_criteria ? data.selection_criteria.length : 0;
+            if (update_of) {
+                $.each(update_of.get('selection_criteria'), function() {
+                    previous_criteria[this.id] = {
+                        'text_en': this['text_en'],
+                        'text_fr': this['text_fr'],
+                        'text_ar': this['text_ar'],
+                    };
+                });
+            }
             for (i = 0; i < criteria_length; i++) {
+                criteria_id = $('#id_selection_criteria-'+i).data('id');
+                criteria_data = previous_criteria[criteria_id] || {};
                 name = data.selection_criteria[i];
                 if (name) {
-                    criteria.push({text_en: name});
+                    criteria_data['text_' + config.get("forever.language")] = name;
+                    criteria.push(criteria_data);
                 }
             }
             data.selection_criteria = criteria;
@@ -150,7 +167,7 @@ module.exports = Backbone.View.extend({
 
             forms.submit($el, 'api/services/', data).then(
                 function success(data) {
-                    window.location = '#/service-list';
+                    window.location = '#/manage/service-list';
                 },
                 function error(missing) {
                     // Missing is a dictionary of errors not already logged on the form
@@ -160,7 +177,8 @@ module.exports = Backbone.View.extend({
 
             return false;
         },
-        "click .form-btn-clear": function() {
+        "click .form-btn-clear": function(e) {
+            e.preventDefault();
             this.$el.find('[name]').each(function() {
                 var $field = $(this);
                 $field.val('');